+++ xteps/src/main/java/com/plugatar/xteps/checked/stepobject/FunctionStep.java
@@ -17,9 +17,9 @@ package com.plugatar.xteps.checked.stepobject;
 
 import com.plugatar.xteps.base.ThrowingFunction;
 import com.plugatar.xteps.base.XtepsException;
-import com.plugatar.xteps.checked.Xteps;
 
-import static com.plugatar.xteps.checked.stepobject.StepObjectsUtils.humanReadableStepNameOfClass;
+import static com.plugatar.xteps.checked.Xteps.stepsChainOf;
+import static com.plugatar.xteps.checked.stepobject.StepObjectsUtils.humanReadableOrEmptyStepName;
 import static com.plugatar.xteps.checked.stepobject.StepObjectsUtils.stepNameWithKeyword;
 
 /**
@@ -39,8 +39,8 @@ public class FunctionStep<T, R, E extends Throwable> implements ThrowingFunction
      *
      * @param step the step
      */
-    protected FunctionStep(final ThrowingFunction<? super T, ? extends R, ? extends E> step) {
-        this.stepName = humanReadableStepNameOfClass(this.getClass());
+    public FunctionStep(final ThrowingFunction<? super T, ? extends R, ? extends E> step) {
+        this.stepName = humanReadableOrEmptyStepName(FunctionStep.class, this.getClass());
         this.stepDescription = "";
         this.step = step;
     }
@@ -93,13 +93,13 @@ public class FunctionStep<T, R, E extends Throwable> implements ThrowingFunction
      * @param t the input argument
      * @return the result
      * @throws XtepsException if Xteps configuration is incorrect
+     *                        or if {@link #stepName} or {@link #stepDescription} or {@link #step} is null
      *                        or if it's impossible to correctly report the step
      * @throws E              if this step threw exception
      */
     @Override
     public final R apply(final T t) throws E {
-        return Xteps.stepsChain().withContext(t)
-            .stepTo(this.stepName, this.stepDescription, this.step);
+        return stepsChainOf(t).stepTo(this.stepName, this.stepDescription, this.step);
     }
 
     /**